+++ backend/hqlib/metric_source/test_report/junit_test_report.py
@@ -107,5 +107,5 @@ class JunitTestReport(test_report.TestReport):
         try:
             return xml.etree.cElementTree.fromstring(contents)
         except xml.etree.cElementTree.ParseError as reason:
-            logging.warning("Couldn't parse report at %s: %s", report_url, reason)
+            logging.error("Couldn't parse report at %s: %s", report_url, reason)
             raise