+++ backend/hqlib/metric_source/test_report/testng_test_report.py
@@ -83,5 +83,5 @@ class TestNGTestReport(test_report.TestReport):
         try:
             return xml.etree.cElementTree.fromstring(contents)
         except xml.etree.cElementTree.ParseError as reason:
-            logging.warning("Couldn't parse report at %s: %s", report_url, reason)
+            logging.error("Couldn't parse report at %s: %s", report_url, reason)
             raise